+++ src/app.tsx
@@ -4,13 +4,20 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { HomePage } from "./pages/home";
 import { AuthCallback } from "./pages/authCallback";
 import { LoginPage } from "./pages/login";
+import { AuthenticatedRoute } from "./utils/authenticatedRoute";
 
 const root = createRoot(document.getElementById("root")!);
 
 const router = createBrowserRouter([
     {
         path: "/",
-        element: <HomePage />,
+        element: <AuthenticatedRoute />,
+        children: [
+            {
+                path: "/",
+                element: <HomePage />,
+            },
+        ],
     },
     {
         path: "/login",
