+++ webpack.config.js
@@ -27,7 +27,11 @@ module.exports = {
             ".cjs": [".cjs", ".cts"],
             ".mjs": [".mjs", ".mts"],
         },
-        fallback: { querystring: require.resolve("querystring-es3") },
+        fallback: {
+            buffer: require.resolve("buffer"),
+            querystring: require.resolve("querystring-es3"),
+            stream: require.resolve("stream-browserify"),
+        },
     },
     plugins: [
         new HtmlWebpackPlugin({
